Show/hide welcome title when search results table is visible.
