+++ src/components/welcomePage/Titel.tsx
@@ -1,65 +1,24 @@
 import { Row, Col } from 'antd';
-import Layout from 'antd/lib/layout/layout';
 import React, { CSSProperties } from 'react'; 
-import TableSearchResults from '../searchPage/TableSearchresults';
-import SearchInput from './SearchInput';
+import { MovieItem } from './welcomeView';
 
-export interface MovieItem {
-    key: string;
-    title: string;
-    imdbScore: string;
-}
-interface State {
-    movieItems?: MovieItem[];
-}
+interface Props {
+    data?: MovieItem[];
+  }
 
-class WelcomeTitle extends React.Component<State> {
-
-    state: State = {
-        movieItems: []
-    }
-
-    handleSearch = (movieItems: MovieItem[]) => {
-        this.setState({ movieItems: movieItems })
-    }
-
-    render() {
+export default function WelcomeTitle(props: Props) { 
+    
+    if (!props.data?.length || false) {    
         return(
-            <Layout style={layoutStyle}> 
-                <Row>
-                    <Col span={24}>
-                        <h1 style={headline}>Welcome to your favourite movie search app!</h1>
-                        <h3 style={tagline}>Search for movies, see the details and rating, add them to your watchlist.</h3> 
-                    </Col>
-                </Row>
-                <Row>
-                    <Col span={24}>
-                        <SearchInput
-                        handleSearch={this.handleSearch}/>
-                    </Col>
-                </Row>
-                <TableSearchResults
-                    data={this.state.movieItems}
-                />
-            </Layout>
+            <Row>
+                <Col span={24}>
+                    <h1 style={headline}>Welcome to your favourite movie search app!</h1>
+                    <h3 style={tagline}>Search for movies, see the details and rating, add them to your watchlist.</h3> 
+                </Col>
+            </Row>
         )
     }
-}
-
-export default WelcomeTitle; 
-
-const layoutStyle: CSSProperties = {
-    backgroundImage: 'url(https://github.com/amandasamuelsson/movie-search-app/blob/master/src/assets/background.png?raw=true)',
-    backgroundSize: 'cover',
-    backgroundRepeat: 'no-repeat',
-    backgroundPosition: 'center center',
-    width: '100%',
-    height: '100vh',
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: '2rem',
-    marginTop: '-6rem'
-    
+    return <div></div>;    
 }
 
 const headline: CSSProperties = {
