+++ src/components/welcomePage/welcomeView.tsx
@@ -1,15 +1,57 @@
-import Layout from 'antd/lib/layout/layout';
-import React from 'react'; 
+import { Layout, Col, Row } from 'antd';
+import React, { CSSProperties } from 'react'; 
+import TableSearchResults from '../searchPage/TableSearchresults';
+import SearchInput from './SearchInput';
 import WelcomeTitle from './Titel';
 
+export interface MovieItem {
+    key: string;
+    title: string;
+    imdbScore: string;
+}
+interface State {
+    movieItems?: MovieItem[];
+}
+class WelcomeView extends React.Component<State> {
 
-function WelcomeView() {
-    return ( 
-    
-    <Layout>
-        <WelcomeTitle />           
-    </Layout>
-        
-    )
+    state: State = {
+        movieItems: []
+    }
+
+    handleSearch = (movieItems: MovieItem[]) => {
+        this.setState({ movieItems: movieItems })
+    }
+
+    render() {
+        return (
+            <Layout style={layoutStyle}>
+                <WelcomeTitle 
+                    data={this.state.movieItems}/>
+                <Row>
+                    <Col span={24}>
+                        <SearchInput
+                            handleSearch={this.handleSearch}
+                        />
+                    </Col>
+                </Row>
+                    <TableSearchResults
+                        data={this.state.movieItems}
+                    />      
+            </Layout>
+    )}
 }
 export default WelcomeView; 
+
+const layoutStyle: CSSProperties = {
+    backgroundImage: 'url(https://github.com/amandasamuelsson/movie-search-app/blob/master/src/assets/background.png?raw=true)',
+    backgroundSize: 'cover',
+    backgroundRepeat: 'no-repeat',
+    backgroundPosition: 'center center',
+    width: '100%',
+    height: '100vh',
+    alignItems: 'center',
+    justifyContent: 'center',
+    padding: '2rem',
+    marginTop: '-6rem'
+    
+}
